chore: move permit.rs into s3/, delete unused key.rs
Permit is an S3-specific type, so it belongs under the s3 module.
key.rs (S3 key encoding) was unused by any consumer.
